Rename DOS audit submodule.
